chore(folders/namespaces): clean up folder structures and namespace spellings/usages
