+++ src/Builders/EmailBuilder.cs
