+++ src/Gateways/S3Gateway.cs
@@ -4,6 +4,12 @@ using Amazon.S3.Model;
 
 namespace form_builder.Gateways
 {
+    public interface IS3Gateway
+    {
+        Task<GetObjectResponse> GetObject(string bucketName, string key);
+        Task<ListObjectsV2Response> ListObjectsV2(string bucketName, string prefix);
+    }
+
     public class S3Gateway : IS3Gateway
     {
         private readonly IAmazonS3 _s3client;
