--- src/Helpers/Session/ISessionHelper.cs
@@ -1,11 +1,0 @@
-﻿namespace form_builder.Helpers.Session
-{
-    public interface ISessionHelper
-    {
-        string GetSessionGuid();
-
-        void RemoveSessionGuid();
-
-        void SetSessionGuid(string value);
-    }
-}
