+++ src/Helpers/Session/SessionHelper.cs
@@ -2,6 +2,15 @@
 
 namespace form_builder.Helpers.Session
 {
+    public interface ISessionHelper
+    {
+        string GetSessionGuid();
+
+        void RemoveSessionGuid();
+
+        void SetSessionGuid(string value);
+    }
+
     public class SessionHelper : ISessionHelper
     {
         private readonly IHttpContextAccessor _httpContextAccessor;
