+++ src/Helpers/ViewRender/ViewRender.cs
@@ -13,6 +13,10 @@ using Microsoft.AspNetCore.Routing;
 
 namespace form_builder.Helpers
 {
+    public interface IViewRender
+    {
+        Task<string> RenderAsync<TModel>(string viewName, TModel model, Dictionary<string, dynamic> viewData = null);
+    }
     public class ViewRender : IViewRender
     {
         private readonly IRazorViewEngine _viewEngine;
