+++ src/Services/BookingService/BookingService.cs
@@ -8,7 +8,7 @@ using form_builder.Models.Elements;
 using form_builder.Providers.Booking;
 using form_builder.Providers.StorageProvider;
 using form_builder.Services.PageService.Entities;
-using form_builder.Utils.Extesions;
+using form_builder.Utils.Extensions;
 using Newtonsoft.Json;
 using StockportGovUK.NetStandard.Models.Booking.Request;
 using StockportGovUK.NetStandard.Models.Booking.Response;
