+++ src/Services/FileUploadService/FileUploadService.cs
@@ -14,6 +14,20 @@ using Newtonsoft.Json;
 
 namespace form_builder.Services.FileUploadService
 {
+    public interface IFileUploadService
+    {
+        Dictionary<string, dynamic> AddFiles(Dictionary<string, dynamic> viewModel, IEnumerable<CustomFormFile> fileUpload);
+
+        Task<ProcessRequestEntity> ProcessFile(
+            Dictionary<string, dynamic> viewModel,
+            Page currentPage,
+            FormSchema baseForm,
+            string guid,
+            string path,
+            IEnumerable<CustomFormFile> files,
+            bool modelStateIsValid);
+    }
+
     public class FileUploadService : IFileUploadService
     {
         private readonly IDistributedCacheWrapper _distributedCache;
