--- src/Services/ValidateService/IValidateService.cs
@@ -1,12 +1,0 @@
-﻿using form_builder.Models;
-using form_builder.Models.Actions;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace form_builder.Services.ValidateService
-{
-    public interface IValidateService
-    {
-        Task Process(List<IAction> actions, FormSchema formSchema, string formName);
-    }
-}
