+++ src/TagParsers/Formatters/TimeOnlyFormatter.cs
@@ -1,10 +1,11 @@
-using form_builder.Utils.Extesions;
+using form_builder.Utils.Extensions;
 
-namespace form_builder.TagParser
+namespace form_builder.TagParsers.Formatters
 {
     public class TimeOnlyFormatter : IFormatter
     {
-        public string FormatterrName { get => "time-only"; }
+        public string FormatterName => "time-only";
+
         public string Parse(string value)
         {
             var dateTime = System.DateTime.Parse(value);
