+++ src/TagParsers/LinkTagParser.cs
@@ -2,8 +2,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
 using form_builder.Models;
+using form_builder.TagParsers.Formatters;
 
-namespace form_builder.TagParser
+namespace form_builder.TagParsers
 {
     public class LinkTagParser : TagParser, ITagParser, ISimpleTagParser
     {
