+++ src/TagParsers/TagParser.cs
@@ -4,8 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 using form_builder.Extensions;
+using form_builder.TagParsers.Formatters;
 
-namespace form_builder.TagParser
+namespace form_builder.TagParsers
 {
     public class TagParser
     {
