+++ src/Utils/ServiceCollectionExtensions/ServiceCollectionExtensions.cs
@@ -44,7 +44,8 @@ using form_builder.Services.RetrieveExternalDataService;
 using form_builder.Services.StreetService;
 using form_builder.Services.SubmitService;
 using form_builder.Services.ValidateService;
-using form_builder.TagParser;
+using form_builder.TagParsers;
+using form_builder.TagParsers.Formatters;
 using form_builder.Validators;
 using form_builder.Workflows;
 using form_builder.Workflows.ActionsWorkflow;
