+++ src/ViewModels/TimeHeaderViewModel.cs
@@ -1,6 +1,6 @@
 using System;
 using form_builder.Enum;
-using form_builder.Utils.Extesions;
+using form_builder.Utils.Extensions;
 
 namespace form_builder.ViewModels
 {
