+++ src/Views/Shared/Time/TimePeriodNoJs.cshtml
@@ -1,4 +1,4 @@
-@using form_builder.Utils.Extesions
+@using form_builder.Utils.Extensions
 @model form_builder.Models.Time.TimePeriod
 @{
     var additionalDivs = Model.Appointments.Count % 3;
