+++ tests/unit-tests/UnitTests/ContentFactory/PageFactoryTests.cs
@@ -4,7 +4,7 @@ using form_builder.ContentFactory;
 using form_builder.Helpers.PageHelpers;
 using form_builder.Models;
 using form_builder.Providers.StorageProvider;
-using form_builder.TagParser;
+using form_builder.TagParsers;
 using form_builder.ViewModels;
 using form_builder_tests.Builders;
 using Moq;
