+++ tests/unit-tests/UnitTests/Mappers/ElementMapperTests.cs
@@ -6,7 +6,7 @@ using form_builder.Enum;
 using form_builder.Mappers;
 using form_builder.Models;
 using form_builder.Providers.StorageProvider;
-using form_builder.Utils.Extesions;
+using form_builder.Utils.Extensions;
 using Moq;
 using Newtonsoft.Json;
 using StockportGovUK.NetStandard.Models.Addresses;
