+++ tests/unit-tests/UnitTests/Services/BookingServiceTests.cs
@@ -22,7 +22,6 @@ using form_builder.Services.PageService.Entities;
 using form_builder_tests.Builders;
 using Microsoft.Extensions.Options;
 using Moq;
-using StockportGovUK.NetStandard.Models.Addresses;
 using StockportGovUK.NetStandard.Models.Booking.Request;
 using StockportGovUK.NetStandard.Models.Booking.Response;
 using Xunit;
