+++ tests/unit-tests/UnitTests/TagParsers/LinkTagParserTests.cs
@@ -3,12 +3,13 @@ using System.Linq;
 using form_builder.Builders;
 using form_builder.Enum;
 using form_builder.Models;
-using form_builder.TagParser;
+using form_builder.TagParsers;
+using form_builder.TagParsers.Formatters;
 using form_builder_tests.Builders;
 using Moq;
 using Xunit;
 
-namespace form_builder_tests.UnitTests.Services
+namespace form_builder_tests.UnitTests.TagParsers
 {
     public class LinkTagParserTests
     {
